Consolidate graphql queries/mutations in one file
