+++ client/src/pages/Transaction/graphql/graphql.ts
@@ -1,0 +1,25 @@
+import { gql } from '@apollo/client';
+
+export const LIST_TRANSACTION_TYPES = gql`
+  query ListTransactionTypes {
+    listTransactionTypes {
+      id
+      name
+      description
+    }
+  }
+`;
+
+export const CREATE_TRANSACTION = gql`
+  mutation CreateTransaction($createTransactionInput: CreateEventInput!) {
+    createEvent(event: $createTransactionInput) {
+      id
+      aggregateId
+      name
+      version
+      data
+      userId
+      createdAt
+    }
+  }
+`;
